pyproject.toml
[project]
name = "health-ai-agent"
version = "0.1.0"
description = "Health AI Agent for medical discharge summary analysis using Med42-v2-8B"
authors = [
    { name = "Rodrigo Castro", email = "rodrigocastropt00@gmail.com" }
]

requires-python = ">=3.11"

dependencies = [
    # Web framework
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    
    # Database
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.23",
    
    # ML/AI
    "huggingface_hub>=0.19.0",
    
    # Data handling
    "pandas>=2.1.0",
    "pyarrow>=14.0.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    
    # Environment & Config
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.1.0",
    
    # Logging & Monitoring
    "structlog>=23.2.0",
    
    # HTTP client (for health checks, etc.)
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",  
    
    # Code quality
    "ruff>=0.1.6",
    "black>=23.11.0",
    "mypy>=1.7.0",
    
    # Development tools
    "pre-commit>=3.6.0",
    "ipython>=8.17.0",
    "jupyter>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["health_ai_agent"]